Config File version tracking and updating
The config files will now track their version, and override an older
version when you update. This will preserve key order but replace
outdated values with the new defaults. If you wish to avoid this, simply
modify the key at the top of your files to the new shuffle move version
and it will think you're already updated.