All stages data updated
All the stage data for special stages is now updated to a consistent
standard. The form of SP_###X where X is an optional uniqueness (or mega
indicating) character, and ### is the dex number. The data now includes
moves and health for all special stages ever ran. Escalation battles use
the highest health of the series.