Implementing Effect threshold handling
The user can now configure the program to only allow effects to trigger
IF and only if they are likely enough to meet a threshold. Threshold of
0 allows all, threshold of 100 only allows "without fail" effects. If
the threshold is exactly equal to the effect chance of an ability for a
combo, the result is undefined due to float imprecision - please use a
value higher or lower than a given odds to ensure your wanted result
will occur.