Portability returns, with more features
You can now have the program launch with a customized save path. The
default is unchanged, but there is now a simple way for any user to
launch the program pointing at a different save location. The portable
windows executable saves to the local directory, and the *.sh and *.bat
scripts allow the use of a path argument to define the save location
(can be relative to java home or absolute).