Fix for issue #60
https://github.com/Loreinator/Shuffle-Move/issues/60

Cause of issue:

Once the number of simulations used for settle checks was increased, the
hash became different for the NumberSpanSimple objects.  I've improved
NumberSpan to include all possible attributes in a constructor, and for
now the problem is solved. I've also changed the equals comparator to be
more forgiving regarding float comparison. NumberSpanSimple has also
been removed, since it is now unneeded.