Non-Support/Support handling expanded
The user can now specify non-support or support by clicking the team
entries in the Team Editor. This is persistent and saved through the
team data. Note: from now on all non-empty teams will have a '-' for the
mega slot if it is not being used. Any old data will work with the new
scheme.

Note: Team data will not be backwards compatible from v0.3.54 onwards
with any version before v0.3.53.